feat: create entity database interface.
In support of developing the new API handler, a persistent storage is required to store entity objects to the server's file system.

The interface of the persistent storage (`EntityDatabase`) is designed in accordance with API handler requirements. Specifically, methods to create, delete, write to, and read from an entity object, as well as to list all stored objects of an Entity type are defined.

Since this change is purely an interface design, unit tests are not included and will be added in future CR of concrete implementations.

Change-Id: I96779c7d499915df138ad124828c0a8cc2701852