# What Implement Entity Database Interface to act as the backend storage system for Crud API using the server's filesystem.
# Why
Require a persistent key-value store to act as the backend storage system to support internal behaviors of the CRUD API.

# How
The implementation uses the server's filesystem to store and retrieve data by ID in directories and provide a persistent key-value store. The interface provides `create_entity`, `delete_entity`, `write_to_entity`, `read_from_entity`, and `list_entities` which corresponds to create, retrieve, update, delete, list actions in the CRUD API.

The unit tests for Entity Database Interface will be added in the later commit. This commit is to streamline the development of the CRUD API.

# Test
Run `cmake ..` and `make` to ensure the code compiles.
The unit tests for Entity Database Interface will be added in the later commit.


Change-Id: Ie119d0c521b1bd6f9a73d8f47793170afea04be7